# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 1,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "ProfileUpdate": {
        "description": "Fields a user may change on their own account.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "language": {
            "anyOf": [
              {
                "enum": [
                  "de",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "language": {
            "default": "en",
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "email",
          "display_name",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "SessionRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "id",
          "provider",
          "created_at",
          "last_seen_at",
          "expires_at",
          "user_agent",
          "current"
        ],
        "title": "SessionRead",
        "type": "object"
      },
      "SetupRequest": {
        "properties": {
          "display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "language": {
            "default": "en",
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "setup_token": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Setup Token",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "setup_token",
          "email",
          "display_name",
          "password"
        ],
        "title": "SetupRequest",
        "type": "object"
      },
...
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "language",
          "timezone",
          "is_active",
          "created_at",
          "last_login_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "admin",
          "user"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ollamail",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Sign in with a local account.",
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Wrong credentials"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too many attempts (rate limit or account lockout)"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "description": "The signed-in user.",
        "operationId": "auth_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "description": "Change display name, language or time zone of the own account.",
        "operationId": "auth_update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Me",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/register": {
      "post": {
        "description": "Create a local account (role ``user``) and sign in, if self-registration is on.",
        "operationId": "auth_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Registration is disabled"
          },
          "409": {
            "description": "E-mail address taken or instance not set up"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too many attempts (rate limit or account lockout)"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sessions": {
      "delete": {
        "description": "Sign out all other sessions, or all sessions with ``include_current=true``.",
        "operationId": "auth_revoke_sessions",
        "parameters": [
          {
            "in": "query",
            "name": "include_current",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Current",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Sessions",
        "tags": [
          "auth"
        ]
      },
      "get": {
        "description": "The own active sessions, most recently used first.",
        "operationId": "auth_list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionRead"
                  },
                  "title": "Response Auth List Sessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/sessions/{session_id}": {
      "delete": {
        "description": "Sign out one of the own sessions.",
        "operationId": "auth_revoke_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in"
          },
          "404": {
            "description": "No such session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/setup": {
      "post": {
        "description": "Create the first admin and sign them in. Only possible while no user exists.",
        "operationId": "setup_create_admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Wrong setup token"
          },
          "409": {
            "description": "The instance is already set up"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin",
        "tags": [
          "setup"
        ]
      }
    },
...
  }
}